moved sortServers function into servers-api
